Add MoneroService examples with detailed, default, and custom binary configurations
- Introduced `monero_ver1_detailed_config.py` for a comprehensive Monero deployment example with specific node roles and configurations.
- Added `monero_ver2_default_config.py` showcasing a concise setup using default parameters.
- Created `monero_ver3_custom_binary.py` to demonstrate the use of user-provided Monero binaries.
- Included necessary custom binary files for the custom binary example.
- Enhanced `MoneroServer.py` to support importing custom binaries into the container.
- Updated `MoneroNodeOptions` to include a field for custom binary imports.
